% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_image_comparisons.R
\name{device_image_comparison}
\alias{device_image_comparison}
\title{Device Images Comparisons}
\usage{
device_image_comparison(df, variable, device_name)
}
\arguments{
\item{df}{Device data-frame in long format.}

\item{variable}{A valid column name denoting the variable of interest.}

\item{device_name}{Device of interest.}
}
\value{
6 heatmaps/scatterplots for each pairwise image grading.
}
\description{
Devices take 3 or 4 images, often with different ratings. This function
visualises the grades between the images for a certain within the data-frame
in either a heatmap or a scatterplot, depending on if the trait measured is
discrete or continuous.
}
\examples{
data(devices, package = "meatrics") # load the data
device_image_comparison(devices, "meat_colour", device_name = "C1")
device_image_comparison(devices, "ema", device_name = "C2")

}
