% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_image_comparisons.R
\name{device_image_diff}
\alias{device_image_diff}
\title{Scatterplots of Device Image Differences (Continuous)}
\usage{
device_image_diff(df, variable)
}
\arguments{
\item{df}{Device data-frame in long format.}

\item{variable}{A valid column name denoting the variable of interest.}
}
\value{
Scatterplots of pairwise image differences, with marks following AMILSC recommended accuracy standards.
}
\description{
Devices take 3 or 4 images, often with different ratings. This function explores the differences between the images for all devices within the data-frame (continuous traits).
For discrete traits, refer to \code{device_image_diff_discrete()}. Note that this function currently only supports MSA-marbling and EMA.
}
\examples{
data(devices, package = "meatrics") # load the data
device_image_diff(devices, "ema")

}
