% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_image_comparisons.R
\name{device_image_diff_discrete}
\alias{device_image_diff_discrete}
\title{Heatmap of Device Image Differences (Discrete)}
\usage{
device_image_diff_discrete(df, variable, device_name)
}
\arguments{
\item{df}{Device data-frame in long format.}

\item{variable}{A valid column name denoting the variable of interest.}

\item{device_name}{Device of interest.}
}
\value{
Up to 6 heatmaps for each different pairwise difference.
}
\description{
Devices take 3 or 4 images, often with different ratings. This function explores the differences between the images for all devices within the data-frame (discrete/categorical traits).
For continuous traits, refer to \code{device_image_diff()}.
}
\examples{
data(devices, package = "meatrics") # load the data
device_image_diff_discrete(devices, "meat_colour", device_name = "C3")

}
