% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{diff_histogram}
\alias{diff_histogram}
\title{Histogram of Grader Differences}
\usage{
diff_histogram(
  df,
  device_status = FALSE,
  variable,
  partitions = 1,
  grader_count = 3,
  bins = 30,
  diff_type = "mean"
)
}
\arguments{
\item{df}{Input dataframe in long format.}

\item{device_status}{A boolean denoting if df is the device data-frame; default assumes grader data-frame.}

\item{variable}{A valid column name denoting the variable of interest.}

\item{partitions}{An integer specifying how many groups to split the variable into.}

\item{grader_count}{An integer specifying how many graders are in the data-frame.}

\item{bins}{An integer specifying what bin-size.}

\item{diff_type}{Either "mean", "median" or "maximum", specifying what kind of difference to take; default is mean.}
}
\value{
A ggplot2 histogram of absolute differences.
}
\description{
This function computes pairwise differences of all graders available in the
data-frame and outputs the result as a histogram. It then summarises these differences
by either taking the mean, median or maximum difference. It allows for partitions in
the variable of interest to be made so that differences can be more clearly
observed along the entire range of the variable.
}
\examples{
data("expert_graders", package = "meatrics")
diff_histogram(expert_graders, variable = "msa_marbling", partitions = 4, grader_count = 3)

}
