% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{extended_bac}
\alias{extended_bac}
\title{Extended BAC Plot}
\usage{
extended_bac(df, device_status = FALSE, variable, grader_count)
}
\arguments{
\item{df}{Input dataframe in long format.}

\item{device_status}{A boolean denoting if df is the device data-frame; default assumes grader data-frame.}

\item{variable}{A valid column name denoting the variable of interest.}

\item{grader_count}{An integer specifying the number of graders there are.}
}
\value{
BAC plot (without LOA)
}
\description{
This function generates an extended Bland-Altman Plot from Möller et al. LOA
is currently not supported in this.
}
\examples{
data("expert_graders", package = "meatrics")
extended_bac(expert_graders, device_status = FALSE, variable = "ema", grader_count = 3)

}
