% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{fleiss}
\alias{fleiss}
\title{Bar chart of category-wise Fleiss Kappa}
\usage{
fleiss(df, device_status = FALSE, df2 = NULL, variable)
}
\arguments{
\item{df}{Input data-frame in long format; must be grader data-frame if df2 is not null.}

\item{device_status}{A boolean denoting if df is the device data-frame; default assumes grader data-frame.
This parameter should be set to FALSE if your df2 is not null.}

\item{df2}{Device data-frame in long format.}

\item{variable}{A valid column name denoting the categorical variable of interest.}
}
\value{
A bar chart of category-wise Fleiss Kappa
}
\description{
Bar chart of category-wise Fleiss Kappa
}
\examples{
data("expert_graders", package = "meatrics")
fleiss(expert_graders, device_status = FALSE, variable = "meat_colour")

}
