% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amilsc_tables.R
\name{human_device_aus_marbling}
\alias{human_device_aus_marbling}
\title{AMILSC Minimum Requirements Table (AUS-Marbling)}
\usage{
human_device_aus_marbling(df, df2, device_name, grader_count = 3)
}
\arguments{
\item{df}{Human grader data-frame in long format}

\item{df2}{Device data-frame in long format}

\item{device_name}{Name of the device that is to be compared with the generalised expert grader.}

\item{grader_count}{An integer specifying the number of expert graders present in df.}
}
\value{
A GT() table following the recommended accuracy standards
}
\description{
This function generates a table following the AMILSC recommended accuracy
standards for cut surface cameras. It computes a generalised expert grader by
taking the median of the devices.
}
