% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{icc_device_comparison}
\alias{icc_device_comparison}
\title{ICC Line Chart over Time (Singular Device v Group of Expert Graders)}
\usage{
icc_device_comparison(df, df2, variable, grader_count)
}
\arguments{
\item{df}{Input expert grader data-frame in long format.}

\item{df2}{Input device data-frame in long format.}

\item{variable}{A valid column name denoting the variable of interest.}

\item{grader_count}{An integer denoting the number of expert graders there are.}
}
\value{
A line chart of ICC.
}
\description{
ICC Line Chart over Time (Singular Device v Group of Expert Graders)
}
\examples{
data("expert_graders", package = "meatrics")
data("devices", package = "meatrics")
icc_device_comparison(expert_graders, devices, variable = "msa_marbling", grader_count = 3)

}
