% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{icc_line}
\alias{icc_line}
\title{ICC3 and ICC3K Line Chart over Time}
\usage{
icc_line(df, df2 = NULL, variable, filtered = NULL)
}
\arguments{
\item{df}{Input dataframe in long format.}

\item{df2}{Input device dataframe in long format.}

\item{variable}{A valid column name denoting the variable of interest.}

\item{filtered}{A character vector of graders you want included in the ICC; can be a mix of both dataframes, If it does not include any grader name from one data-frame, all graders from that data-frame will be used.}
}
\value{
A line chart of ICC3 and ICC3k with error bars.
}
\description{
ICC3 and ICC3K Line Chart over Time
}
\examples{
data("expert_graders", package = "meatrics")
data("devices", package = "meatrics")
icc_line(expert_graders, df2 = devices, variable = "msa_marbling")

}
