% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{merge_device_grader_long}
\alias{merge_device_grader_long}
\title{Merges 2 data-frames (device and grader) into long format}
\usage{
merge_device_grader_long(df1, df2, variable)
}
\arguments{
\item{df1}{Grader data-frame in long format.}

\item{df2}{Device data-frame in long format.}

\item{variable}{A valid column name denoting the variable of interest.}
}
\value{
A merged long data-frame of device and grader data, based on common body_no and kill_date combinations.
}
\description{
Merges 2 data-frames (device and grader) into long format
}
