% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{pairwise_ccc}
\alias{pairwise_ccc}
\title{Pairwise CCC Radar Graph}
\usage{
pairwise_ccc(
  df,
  device_status = FALSE,
  df2 = NULL,
  variable,
  partitions = 1,
  message = FALSE
)
}
\arguments{
\item{df}{Input data-frame in long format.}

\item{device_status}{A boolean denoting if df is the device data-frame; default assumes grader data-frame.
This parameter is arbitrary if your df2 is not null.}

\item{df2}{Device data-frame in long format.}

\item{variable}{A valid column name denoting the variable of interest.}

\item{partitions}{An integer specifying how many groups to split the variable into.}

\item{message}{A boolean denoting if the CCC values should be printed to terminal.}
}
\value{
A radar chart of pairwise CCC.
}
\description{
This function computes pairwise CCC for all graders present in the data-frame
and plots it on a radar chart. It allows for partitions in the variable of
interest to be made so that differences can be more clearly observed along the
entire range of the variable. If two data-frames are input, this function will
assume that the first data-frame belongs to the expert graders and will compute
the mean of the expert graders' ratings, creating a representative expert grader to compare
against each device.
}
\examples{
data("expert_graders", package = "meatrics")
pairwise_ccc(expert_graders, device_status = FALSE, variable = "ema", partitions = 5)

data("devices", package = "meatrics")
pairwise_ccc(expert_graders, device_status = FALSE, df2 = devices, variable = "msa_marbling", partitions = 5)

}
