% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{pairwise_diff}
\alias{pairwise_diff}
\title{Helper function of pairwise_stats()}
\usage{
pairwise_diff(df, device_status = FALSE, variable)
}
\arguments{
\item{df}{Input data-frame in long format}

\item{device_status}{A boolean denoting if this is the device data-frame; default is human grader}

\item{variable}{A valid column name denoting the variable of interest}
}
\value{
A table containing pairwise maximum difference, mean difference and median difference
}
\description{
Helper function of pairwise_stats()
}
