% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{pairwise_heatmap}
\alias{pairwise_heatmap}
\title{Pairwise Heatmaps (Discrete Traits)}
\usage{
pairwise_heatmap(df, device_status = FALSE, variable)
}
\arguments{
\item{df}{Data-frame in long format.}

\item{device_status}{A boolean denoting if df is the device data-frame; default assumes grader data-frame.
This parameter is arbitrary if your df2 is not null.}

\item{variable}{A valid column name denoting the variable of interest}
}
\value{
3 pairwise heatmaps of graders' assessment on the discrete variable of interest
}
\description{
This function plots pairwise heatmaps of all graders in the input
data-frame on any discrete variable of interest.
}
\examples{
data("expert_graders", package = "meatrics")
pairwise_heatmap(expert_graders, device_status = FALSE, variable = "aus_marbling")

}
