% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{pairwise_scatter}
\alias{pairwise_scatter}
\title{Pairwise Scatterplots}
\usage{
pairwise_scatter(df, device_status = FALSE, df2 = NULL, variable, line = TRUE)
}
\arguments{
\item{df}{Data-frame in long format.}

\item{device_status}{A boolean denoting if df is the device data-frame; default assumes grader data-frame.
This parameter is arbitrary if your df2 is not null.}

\item{df2}{If not null, df2 must be the device data-frame in long format.}

\item{variable}{A valid column name denoting the variable of interest}

\item{line}{A boolean for plotting y = x; default is set to TRUE.}
}
\value{
A pairwise scatterplot of all graders' assessment on the variable of interest
}
\description{
This function plots pairwise scatter-plots of all graders in the input
data-frame on any variable of interest.
}
\examples{
data("expert_graders", package = "meatrics")
pairwise_scatter(expert_graders, variable = "ema", line = TRUE, device = FALSE)

pairwise_scatter(expert_graders, variable = "msa_marbling", line = FALSE, device = FALSE)

}
