% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{pairwise_stats}
\alias{pairwise_stats}
\title{Pairwise Statistics Table (continuous traits only!)}
\usage{
pairwise_stats(df, device_status = FALSE, variable, by_date = FALSE)
}
\arguments{
\item{df}{Input data-frame in long format}

\item{device_status}{A boolean denoting if this is the device data-frame;
default assumes grader data-frame.}

\item{variable}{A valid column name denoting the variable of interest.}

\item{by_date}{A boolean denoting if the statistics should be split by date.}
}
\value{
A table containing pairwise maximum difference, mean difference,
median difference, R-squared, Kendall's Tau and Spearman's Rho.
}
\description{
Pairwise Statistics Table (continuous traits only!)
}
