% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{to_wide}
\alias{to_wide}
\title{Wide-formatted Data-frame}
\usage{
to_wide(df, variable, device = FALSE, device_proc = "mean")
}
\arguments{
\item{df}{Input data-frame in long format.}

\item{variable}{A valid column name denoting the variable of interest.}

\item{device}{A boolean denoting if this is the device data-frame; default assumes grader data-frame.}

\item{device_proc}{Summarises how the multiple device images should be dealt with. If it is a continuous variable,
you can specify to take either the mean or median as the value. Otherwise, specify an integer denoting the image number to take.}
}
\value{
A wide data-frame of the variable of interest with graders as column names.
}
\description{
Wide-formatted Data-frame
}
