% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{threshold_diff}
\alias{threshold_diff}
\title{Table containing counts of exact agreement and ranges of input thresholds}
\usage{
threshold_diff(
  df,
  device_status = FALSE,
  variable,
  grader_a,
  grader_b,
  units,
  visuals = "table",
  partitions = 1
)
}
\arguments{
\item{df}{Input dataframe in long format}

\item{variable}{A valid column name denoting the variable of interest}

\item{grader_a}{First grader}

\item{grader_b}{Second grader}

\item{units}{Specified units}

\item{visuals}{Will output either a stacked bar chart or a table. Can be either "table" or "bar", default is "table".}

\item{partitions}{An integer specifying how many groups to split the variable into}
}
\value{
A GT() table or ggplot2 stacked bar chart, categorizing differences into 4 categories.
}
\description{
Table containing counts of exact agreement and ranges of input thresholds
}
